feat: Remove v1 reconcile and initial refactor of reconcile_v2.go
