+++ internal/controller/reconciler/clickhouse.go
@@ -1,4 +1,4 @@
-package v2
+package reconciler
 
 import (
 	"context"
