+++ internal/controller/reconciler/manifest_order.go
@@ -1,4 +1,4 @@
-package v2
+package reconciler
 
 import (
 	"maps"
