+++ internal/controller/reconciler/networking_route_builders_test.go
@@ -1,4 +1,4 @@
-package v2
+package reconciler
 
 import (
 	. "github.com/onsi/ginkgo/v2"
