+++ internal/controller/reconciler/telemetry_chart_test.go
@@ -1,4 +1,4 @@
-package v2
+package reconciler
 
 import (
 	"os/exec"
