+++ internal/controller/reconciler/v2_suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v2_test
+package reconciler_test
 
 import (
 	"testing"
